+++ backend/src/main/java/de/thu/thutorium/database/dbObjects/UniversityDBO.java
@@ -29,7 +29,7 @@ public class UniversityDBO {
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   @Column(name = "university_id")
   @Setter(AccessLevel.NONE)
-  private Integer universityId;
+  private Long universityId;
 
   /**
    * The name of the university. This field is required and has a maximum length of 255 characters.
@@ -56,7 +56,7 @@ public class UniversityDBO {
   private List<AffiliationDBO> affiliations;
 
   /**
-   * Constructs a UniversityDBO with an empty set of affiliations..
+   * Constructs a UniversityDBO with an empty set of affiliations.
    */
   public UniversityDBO() {
     this.affiliations = new ArrayList<>();
